added 'buttons' for my book genres
I have continued to work through more basic additions of HTML and CSS. I added buttons for the book genres I want to write about (I can't click on them yet, but they are there) and added a box around the title of my page. In class today I want to play around with the color scheme of my page and work on adding more content to it as well.