Final draft
Added in all my photos and book summary's. As hard as I tried to watch videos and read I couldn't figure out how to put my tabs in a row, but other than that I am really pleased with how it turned out. I still have a lot to learn with this, but for my first time I am really excited about how it came out.